--- models/product.py
@@ -1,10 +1,0 @@
-from peewee import Model, CharField, DecimalField
-from .db import db
-
-class Product(Model):
-    name = CharField()
-    attribute = CharField()
-    atackpower = DecimalField()
-
-    class Meta:
-        database = db
