--- templates/product_edit.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-    <meta charset="UTF-8">
-    <title>武器編集</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='base-style.css') }}">
-</head>
-<body>
-    <h1>武器を編集</h1>
-    <ul class="breadcrumb">
-        <li><a href="{{ url_for('index') }}">HOME</a></li>
-        <li><a href="{{ url_for('product.list') }}">武器一覧</a></li>
-        <li>武器編集</li>
-    </ul>
-
-    <form action="{{ url_for('product.edit', product_id=product.id) }}" method="post">
-        <label>武器名: <input type="text" name="name" value="{{ product.name }}" required></label><br>
-        <label>攻撃力: <input type="number" step="1" name="atackpower" value="{{ product.atackpower }}" required></label><br>
-        <label>武器タイプ: <input type="text" name="attribute" value="{{ product.attribute }}" required></label><br>
-        <button type="submit">保存</button>
-    </form>
-</body>
-</html>
